Split the organization feature settings out of the form component
The "Settings for organization features" card was over half of
Admin::OrganizationForm::Component and owned most of its private
methods. It becomes Admin::OrganizationFeatureSettings::Component, a
sibling like the location fields, taking the form component from 368
lines to 208.

Along the way: search_radius_attribute joins its stolen-message
counterpart as a method rather than an inline ternary, and the custom
reg label loop guards with next instead of a nil sentinel. The rendered
HTML is unchanged.
